refactor: 상품 관련 폴더 추가 및 파일 이동
