+++ lib/product/component/bottom_navigation_widget.dart
