+++ lib/product/component/product_detail_widget.dart
