+++ lib/product/component/product_info_widget.dart
