+++ lib/product/component/seller_info_widget.dart
