+++ lib/product/view/declaration_dialog.dart
